upipe-hbrmt: change pixel argument types to uintptr_t
The reasoning for this is:
* We want it to be unsigned because we don't expect a negative value to
  be passed.  We perform unsigned shifts on the value in some cases.
* We want it to be register sized so that we do not have to extend it on
  x86-64.

+++ lib/upipe-hbrmt/sdidec.c
@@ -21,7 +21,7 @@
 #include <stdint.h>
 #include "sdidec.h"
 
-void upipe_sdi_to_uyvy_c(const uint8_t *src, uint16_t *y, int64_t pixels)
+void upipe_sdi_to_uyvy_c(const uint8_t *src, uint16_t *y, uintptr_t pixels)
 {
     pixels *= 2; /* change to number of samples */
     for (int i = 0; i < pixels; i += 4) {
